Moved media picker to its own package
