+++ WooCommerce/src/main/res/layout/fragment_wpmedia_picker.xml
@@ -5,7 +5,7 @@
     android:background="@color/black"
     android:orientation="vertical">
 
-    <com.woocommerce.android.widgets.WPMediaLibraryImageGalleryView
+    <com.woocommerce.android.ui.products.wpmediapicker.WPMediaLibraryGalleryView
         android:id="@+id/wpMediaGallery"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
